+++ app/server/api/username/validate/[username].get.ts
@@ -5,7 +5,9 @@ import { unauthorized } from '~/utils/nuxt'
 import { usernameSchema, usernames } from '~~/db/schema'
 
 export default eventHandler(async (event) => {
-  const { username } = await getValidated(event, 'params', z.object({ username: z.string() }))
+  const { username } = await getValidated(event, 'params', z.object({
+    username: z.string(),
+  }))
 
   const orm = event.context.orm
   const user = event.context.user